store: cache the series-id written to store to dedupe writes
We're writing the series label index to bigtable for every chunk
We now cache the series-id and write only if we didn't write it before

Signed-off-by: Goutham Veeramachaneni <gouthamve@gmail.com>

--------

This is a squashed commit but only including Tom's commits'
description for attribution.

Review feedback.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>

Write back cache keys after they have be written to store.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>