promtail: support non-integer position offsets in positions file
The easiest way to consistently refer to a journal entry when seeking in
the systemd journal is to use the cursor, which is an aribtrary opaque string
used by systemd to refer to an entry. The positions file used to only
store int64 values and has been updated to store generic strings.

Old calls to the positions.Positions.Get have been updated to accept an
error, which may fail if a non-int value was written for a file.
GetString and PutString have been added to put a raw string value.

This change is backwards compatible with existing positions files;
yaml.v2 will read in the existing positions as strings. When the
positions file is saved to disk, the values will be updated with
quotes around them.