promtail: remove journal Since field and refactor Target interface
The journal Since field has been removed; it's unlikely to be used as
long as the positions storage works and its behavior was confusing (as
it specified an offset from initialization time of the JournalTarget,
which could be unpredicable).

The Target interface has been refactored to remove the unecessary
DiscoveredLabels and Labels methods. Only Labels was being used to
determine if the Target was dropped; this has been changed to check to
see if a Target's type was listed as dropped.