chore(packaging): master tagging
Contrary to what has been done before, this reverts to something close to the
old behavior:

If a git tag is built, the name of the tag is used as the docker tag.
If the master branch is built, it is tagged as master-SHA, master and latest

If a regular branch is built, it is not pushed.