promtail: use debian as base image to support systemd journal reading
For promtail to be able to read from the systemd journal, libsystemd.so
must be present at runtime. There is currently no package for alpine
that provides this file and such, debian should be used instead. A
future solution for providing an alpine-based build should be
considered in the future and will be tracked in #838.

This commit also updates the Dockerfile to default to using the
latest loki-build-image. If a specific version is required, it should
be provided as a build arg to docker build.