chore(ci/cd): build containers using drone.io (#891)
Because drone.io provides us with native ARM runners, we evaluate the platform to figure out if it is a suitable replacement for CircleCI and qemu.

Configuration is done using .drone/drone.jsonnet, the .drone/drone.yml is generated from that.
Images are built on amd64, arm64 and arm, natively and without emulation. This allows to get rid of cross-compilation, which in turn allows more advanced builds (CGO, etc.)

Hopefully this will make the CI faster and more reliable in the future. Let's see!