chore(packaging): set tag length to 7
relied on git approximation before, which causes problems with other parts of
the automation.

Because we do not need the tag to be strictly unique, but just to identify the
latest version, limiting to 7 is fine.

If this ever causes problems, raise the limit but make sure to update all other
places as well.