promtail: add "cutoff" field to configure cutoff for journal reading
The journal scrape config in promtail has been updated to support
a "cutoff" field. cuttoff determines the oldest journal entry
promtail will read when starting the journal reader. When
unspecified, cutoff defaults to 7h.

Even if a position in the journal is saved in the promtail
positions file, that position may be ignored if the entry corresponding
to that position is older than the cutoff time.

Example promtail config for getting up to the last 12h of journal
entries:

  server:
    http_listen_port: 9080
    grpc_listen_port: 0

  positions:
    filename: /tmp/positions.yaml

  clients:
    - url: http://localhost:3100/api/prom/push

  scrape_configs:
  - job_name: journal
    journal:
      cutoff: 12h
      path: /var/log/journal
      labels:
        job: systemd-journal
    relabel_configs:
      - source_labels: ['__journal__systemd_unit']
        target_label: 'unit'