promtail: add "max_age" field to configure cutoff for journal reading (#921)
The journal scrape config in promtail has been updated to support
a "max_age" field. max_age determines the oldest journal entry
promtail will read when starting the journal reader. When
unspecified, max_age defaults to 7h.

Even if a position in the journal is saved in the promtail
positions file, that position may be ignored if the entry corresponding
to that position is older than the max_age cutoff time.

Example promtail config for getting up to the last 12h of journal
entries:

  server:
    http_listen_port: 9080
    grpc_listen_port: 0

  positions:
    filename: /tmp/positions.yaml

  clients:
    - url: http://localhost:3100/api/prom/push

  scrape_configs:
  - job_name: journal
    journal:
      max_age: 12h
      path: /var/log/journal
      labels:
        job: systemd-journal
    relabel_configs:
      - source_labels: ['__journal__systemd_unit']
        target_label: 'unit'