chore(ci/cd): breaking the circle
As the trial period for drone.io went really well (not a single failed build,
apart from merge-failures), I think it is time to make the switch to the new,
native ARM and much more performant platform.

This should fix all kinds of issues and be a solid architecture for the future
to level up our CI/CD in terms of performance and reliability.