fix(pkg/cfg): distinguish set values from defaults
Previously, the flag functionality relied on comparing archived defaults with
the merged data after flag.Parse(). This inevitably led to a bug, that this
package was able to decide whether a value is on default, or set by user to the
same as default. This meant the user value was always lost.

This now switches it to a more traditional usage of the flag package:

1. flag.***Var already sets the default value on the pointer
2. intermediate sources can make changes here
3. flag.Parse() ONLY sets new values