build: override promtail release build with cgo-enabled binary
The published binary of promtail for linux/amd64 was not built with CGO
support, so journal tailing could not be used.

This is kind of a dirty hack: we just re-build promtail at the end of
building everything else with CGO enabled. It's not clear how it could
be done more cleanly.

Fixes #1057.