helm: update default terminationGracePeriodSeconds to 4800
Users that run Loki using the Helm chart may not modify the value for
terminationGracePeriodSeconds that allows Loki enough time to flush all
of its data. If Loki is forcefully terminated by Kubernetes during a
flush, missing data or an unhealthy ingester in the ring in
microservices mode can occur.

This commit changes the grace period to 4800 seconds, the value used in
the Jsonnet code.