pkg/loki: unmarshal module name from YAML
The module name field in the config could not be unmarshalled from a
string, causing parsing the config to crash.

This commit also changes the error message when the config could not be
parsed to print directly to stderr, as the logger won't be initialized
at the point where parsing the config fails, leading to no output in the
console.

Fixes #1259.