fluent-bit: tenant ID configuration
Both promtail and the fluentd plugin already have support for sending a
tenant ID, and all necessary for doing so in fluent-bit is actually
passing through the configuration item. Configuration is aligned with
the fluentd plugin for consistency.

We're lucky, since no configuration will result in an empty string that
will be treated as no tenant ID by the Loki shipping client.

Signed-off-by: Jens Erat <email@jenserat.de>