pkg/promtail: remove journal target forced path (#1298)
* pkg/promtail: remove journal target forced path

The forced default path for the journal target (/var/log/journal) did
not apply to all Linux systems: systemd actually defaults to reading
from _two_ default paths: /var/log/journal and /run/log/journal. Not
setting an explicit default will enable both of those paths to be
searched for journal logs.

The comment mentioning the local machine ID was technically accurate,
but is normally dealt with by bind mounting /etc/machine-id. That's the
standard approach for running fluent-bit in a container, and should be
the standard approach for Promtail as well.

Fixes #1286.