pkg/logentry: refactor stages to use callbacks
Proof of concept, do not merge!

This commit introduces a ChanedStage interface to control when or if the
next stage in the pipeline gets ran. Existing stages have been wrapped
as chained stages.

Note that a temporary "delayStage" has been added to prove that there
are issues with out of order logs when using this mechanism. This may be
due to a bug, but I'm not sure yet.