production/helm: Support taking Loki configuration as a string
Partial fix for #1392

Currently .Values.config is a map, meaning setting values only
merges/overrides existing values, with no way to override the whole
configuration without the merging behavior.

To support supplying the configuration as a string, I've added a new
option `configString` that takes precedence over `config`, for the
contents of the `loki.yaml` configuration.