production/helm: Remove previously added configString and handle config correctly as a dict and string
Uses Go templating with sprig reflection functions to determine if
Loki's .Values.config is a string, and handles it correctly as a string
if so, or reverts to existing dictionary style behavior with the users
config being merged with the default configuration.

Full fix for #1392, as it's both backwards compatible, and requires no
new fields.