promtail: add support for passing through journal entries as JSON
This commit adds an extra configuration field (`json`) to the journal
target configuration schema. When set to true, messages from the journal
target will be passed through as JSON, containing all of the fields from
the entry as keys and their values as the value.

For backwards compatibility, the value for `json` defaults to false.

Closes #1288.