Use original README.md for fluentd plugin explanations
`doc/client/fluentd.m` includes obsolete explanations and it confuses new users.
For example, label_keys option was removed at https://github.com/grafana/loki/pull/1186
but `fluentd.md` has the explanation.

It's simple to maitain `fluentd/fluent-plugin-grafana-loki/README.md` for
the explantion of the output plugin and use the link when we share how to use it.