Fixes possible panic in json pipeline stage.
This was reported to me today and it happens because the previous source can be nil.
This was only happening when debug log was on. And the problem was only in the json because it was using .String on a nil value.

I've added a test and activate debug during tests to cover those path too.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>