Add no-file and keep-file option.
By default no-file is false, this means the driver will create json log files that can be used to run `docker logs` command.
If no-file is true, no files are created and so `docker log` won't work. This allow the ability to run the driver without caring of space left on the host.

keep-file tells the driver to keep file even after the container has been stopped, by default this options is false. This means files are cleaned up once containers are stopped and not history is left to be used for `docker logs`.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>