Accept suffixes to pod name in instance labels
Currently, the instance label is just the pod name, but that's
insufficient,
cf. https://github.com/grafana/deployment_tools/issues/3441 .

In the near future, we'll suffix the pod name in the instance label
with container name and/or port name. To make this mixin still work,
adjust the `instance` label matcher.

Signed-off-by: beorn7 <beorn@grafana.com>