Fixes iterators boundaries.
The start should be inclusive and the end is supposed to be exclusive.
If start and end are equal, then only one entry can be returned which need to also be equal to start.

Fixes #2124

Those issues were edge cases where the boundaries of a block or iterator would be equal to the start.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>