add RegisterFlagsWithPrefix to config structs
All config structs now have a RegisterFlagsWithPrefix function that
allows specifying a prefixed string to append to all flag names. This
may be used by clients to namespace support for Promtail in a larger
application (e.g., grafana/agent).

When implementing this PR, I noticed a few config structs were
incorrectly registering to the global flagset rather than the flagset
passed to RegisterFlags. This PR also includes a fix for that and
ensures every option is registered to the flagset in the parameter.

Related to #2405.