Collapse multiline logs based on a start line. (#3024)
* Collapse multiline logs based on a start line.

Summary:
This is a very simple approach based on #1380 to provide multiline
or block log entries in promtail.

A `multiline` stage is added to pipelines. This stages matches a start
line. Once a start line is matched all following lines are appended
to an entry and not passed on to downstream stages. Once a new start
line is matched the former block of multilines is sent.

If now new line arrives withing `max_wait_time` the block is flushed to
the next stage and a new block is started.

* Test multiline stage process.

* Format code.

* Flush multiline block after `max_lines`.

* Capture internal state of the stage.

* Process different multiline streams in parallel.

* Start documenting multiline stage.

* Give an example configuration for `multiline` stage.

* Make linter happy.

* Pass through entries until first start line.

* Update pkg/logentry/stages/multiline.go