[POC] last_over_time and first_over_time.
This is an attempt to allow users to select the first or last discrete sample values from a range vector.
This way we can create gauge, the only drawback being the log need to exists at least in every range.

May be we could transfer the last or first  value to the next range, but it will still be a problem for small ranges.

Still wondering how useful this is compare to min/max_over_time.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>