Guarantee unique timestamps even if there are unparseable log-lines in between
The previous implementation checked for strict timestamp equality and required,
keeping track of the current nanosecond offset. It also had a major flaw where
an unparseable line in between two log lines with the same timestamp would mean
uniqueness wouldn't be guaranteed, and the logs would potentially get out of
order. This implementation solves this by checking if the new timestamp is
equal to or before the last known timestamp and if so sets it to one ns after
the last known timestamp. This works well seeing that fudge on failure sets
the last known timestamp (and uses the same logic for multiple failed lines)