Fixes head chunk iterator direction.
For backward queries, since LogQL parser we are using a heapIterator in the headchunk to re-order properly entries.
But we also reverse all iterators in the memchunk code, even the headchunk which causes reversal of already reversed entries.

This PR skips reversal of the headchunk.

Fixes #3345
Fixes #3208

This has for side effects:
-  when using replication you would not dedupe data properly anymore, since the data is not correctly ordered accross batches.
-  when using limit it would miss entries.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>