Fixes a race in the scheduling limits.
Found out the hard way that sharding might not listen correctly to context and so still run queries
after the upstream request ends. This can lead to panic when scheduling work which send on a closed channel.

I've added a safe guard to wait for all jobs to end before ending the original request but also attemped to fix the sharding downstreamer
to properly stop sending work when context is closed.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>