Avoid to mutate log line when unpacking.
We recently introduce this, the headchunk uses unsafe string to avoid allocations when processing log line, this means we should never mutates log line in pipeline.
We can reslice though but the problem is that we still need to remove `\\` from the content.
Since we had to pay for a copy, I also improved the parsing to negate the hit.

This needs a bigger refactoring to avoid this mistake again.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>