Properly release the ticker in Loki client.
I think this was only impacting the docker driver who would start/stop for each new followed containers.
My assumption is that this would slowly build up and eat CPU over time.

Fixes #3319

I arrived to this conclusion since strace was showing a crazy amount of futex syscall and nothing else seems to be the cause.

:pray:

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>