Improve retention mark files.
This PR is two fold:

1) Rotate mark files when they reached 100k marks, this is to ensure we max out marks file
and don't create file that may be too big.

2) Instead of inserting marks using the chunk id as the key, we insert using a natural sequence.
This has two benefits:
- Keys are ordered, and so insertion are faster.
- Allows us to use boltdb fill percent to 100%, this means boltdb won't over allocate for inserting key in between unordered data.

Why ?

I realize that when inserting huge amount of marks, this operation can take up to hours since boltdb has to re-allocate pages over and over.
This is mainly because chunkid arrive without  specific order.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>