Fixes a goroutine leak in the store when doing cancellation.
Found out that we might have no listener and trying to send the information of cancellation.
I've moved the error information down the respective log and sample iterators so that we don't need to send
it via the channel.

This seems to have cause some pretty big memory leak on our side.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>