Update out_loki.rb
This PR allows loki to receive the namespace labels set by the kubernetes_metadata plugin.
As said in comment, I'm not happy the long name for the label. For example, it gives me "namespace_label_kustomize_toolkit_fluxcd_io_checksum" as label name. Feel free to find a shorter name.