fix: Restore /config endpoint and correct handlerFunc for buildinfo
As far as I can tell from reading the "next" documentation, the
/loki/api/v1/status/buildinfo endpoint should return the build
information. The /version endpoint is also listed further down in the
documentation but from the code prior to this change, it looks like it
should perform the same function as buildinfo endpoint so I have
removed it.

I've updated the documentation to reflect my changes.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>