Update to the Loki tanka install docs to reflect changes made in tanka.
Also pulled up the docs piece about the bolt-db-shipper-shared-store so users know they must modify it, rather than seeing it as a note further down.

I think we're probably still missing one more update to the yaml to add the `schema_config` override, but that can come as a separate pr.