cmd/logcli: add --follow flag as an alias for --tail
When using 'logcli query --tail', '--tail' behaves similar to the 'tail'
command, which uses '--follow' flag, if you want to continuously
follow the appended data to the file.

I believe '--follow' flag is more natural for system administrators to
use rather than '--tail' if one wants to "follow" the incoming logs, so
this commit adds one, as an alias for '--tail'.

Closes #3570

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>