docs: fix link to Promtail documentation (#4293)
* chore: fix link to Promtail documentation

The current link to Promtail installation documentation in the file leads to 404 error. I observed that the `../docs/clients/` folder was refactored into `../docs/sources/clients` at same time in past. However, the `loki/docs/sources/clients/promtail/installation.md` file is not very convenient if we talk about UX. So, I've replaced the wrong link with the one to rendered beautiful documentation

* chore: make link relative