Fixes a bug in the block cache code.
Because chunks can overlap and that can causes multiple processing of the same chunk within batches, we decided to cache the
evaluation of it.

However there was a bug here causing incorrect query result, specially for metric queries. The bug would occur when we would only partially consume a block (overlaping block between multiple batches) and then try to replay it.
In which case it wouldn't actually replay it but keep consuming the underlaying iterator.

This took us 2 days to figure out, thank Danny for helping along.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>