Docs: correctly represent product name (#4416)
The product name is Grafana Loki.  The first use of the product name on
each web page should be the full product name.  After that, within each
page, it is fine to use just Loki.

This PR also set up some frontmatter with weights where it was missing
on a few files, and correct where "Loki" was not capitalized.