docker-driver: Add a warning
This really caused me a lot of pain. I haven't done very scientific tests, but I did 
need to restart docker a few times, for it to hang, to then forcibly killing it and/or
some of its containers, and then I finally just removed this configuration.

I now use a logging override in my docker compose file.