Revert distributor defaulting to inmemory. (#4638)
- [PR 4440](https://github.com/grafana/loki/pull/4440/files) modified
  the distributor setup to use inmemory by default (instead of consul).
That was necessary because, by changing the default rate limiting to
global, a ring was necessary, and since by default the ring was set to
consul, the project couldn't run with the default configs. That's not
the case anymore: right now we are reusing the same config set for the
ingester (which users already have properly configured otherwise they
wouldn't be able to run the project). List of changes:
  - Remove changelog entries
  - Remove upgrading guide entries
  - Remove added tests (unnecessary checks)
  - Remove value overriding