Fixes Promtail User-Agent.
This was because promtail client was still user the old version.Version which is initialized at run-time via init.

But if we use it also at init time there's no guarantee that the version will be set.

So intead I'm using the original version variable stamped in the binary and not at runtime.

Fixes #4686

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>