Refactor NewMemcachedClient constructor
Previously NewMemcachedClient was returning the MemcachedClient
interface that only had a subset of cache methods. This was
limiting the use of the constructor to a single type of use.

The changes introduced in this PR are simply:

- Rename the MemcachedClient interface into MemcachedBasicClient
- Rename memcachedClient type to MemcachedClient
- Now the constructor returns the type instead of the interface
- Also MemcachedClient now does not anonymously export memcache.Client
  methods and sync.Mutex methods

This enables the the constructor to be used for other purposes where the
caller may want to use other available methods in the memcached client
like Add, Delete, CompareAndSwap etcetera.