Support Redis Cluster Configuration Endpoint
The go-redis options requires multiple redis addresses for a cluster.
In some scenarios like AWS ElastiCache, a single configuration endpoint
is used to resolve all cluster nodes.

This change adds an extra step to resolve all nodes and pass them to go-redis.
It remove the need to specify multiple cluster nodes in the configuration.

No tests were added for this scenario since it requires patching
the Go resolver during runtime.

Signed-off-by: Siavash Safi <siavash.safi@gmail.com>