Remove sync() call in FileTarget constructor
The constructor `NewFileTarget` is called from within a locked code
block by the FileTargetManager. However, the constructor also calls
`sync()` which starts and stops watching new directories. These methods
`startWatching` and `stopWatching` both send to the `targetEventHandler`
channel, causing the `FileTargetManager` to receive the event in a
goroutine (`watch()` function) running in the background.
Handling the event calls `sendFileCreateEvent` on the syncers, which
subsequently tries to write to the `fileEventWatcher` channel which is
never read because of the locked mutex.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>