Hint about potential high cardinality metrics
Add warning about potential high cardinality metrics due to the file path being included in some of the metrics. 
Depending on the setup, such config can result in a very large amount of label values, which is not recommended by Prometheus.
This can result in increased storage, slow down of queries, extra costs and so on.

Address what has been discussed in #5553.