Support limits in multi-tenant queries. (#5626)
**What this PR does / why we need it**:
This changes adds support for the the following limits in multi-tenant queries:
- `MaxQuerySeries`
- `QuerySplitDuration`
- `MaxEntriesLimitPerQuery`
- `MinShardingLookback`
- `MaxQueryLength`
- `MaxQueryParallelism`
- `MaxQueriersPerUser`
- `MaxCacheFreshness`


It always uses the more restricted limit. That means the smallest `MaxQuerySeries` but biggest `QuerySplitDuration`. The split duration is special. Zero means disabled.

Also, the global default resolver is the multi tenant resolver when multi-tenant queries are enabled. This will result HTTP 400 errors in case the user sends multiple tenants to endpoints that do not support it.

**Which issue(s) this PR fixes**:

**Special notes for your reviewer**:

**Checklist**
- [ ] Documentation added
- [x] Tests updated
- [x] Add an entry in the `CHANGELOG.md` about the changes.

Co-authored-by: JordanRushing <rushing.jordan@gmail.com>