Update Fluent Bit docs (#5402)
* Change links to fluentbit.io website

The documentation for the Fluent Bit Loki output plugin references the
fluentbit.io website multiple times. However, the structure of this
website seems to have changed. This commit updates the links to the
correct URLs (if possible).

Additionally, renamed occurences of "fluentbit" to "Fluent Bit", which
seems to be the correct spelling of the product according to their
website.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>

* Remove trailing commas from JSON code blocks

Trailing commands are syntactically not allowed in JSON.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>

* Apply suggestions from code review

Co-authored-by: Karen Miller <84039272+KMiller-Grafana@users.noreply.github.com>

Co-authored-by: Karen Miller <84039272+KMiller-Grafana@users.noreply.github.com>
(cherry picked from commit 1a7b170cdbda27e17719a22a900136e5caf6ecef)