Trigger build for Loki build image version 0.20.3
In commit 435391661f4e0f03eb378f77c07aaad5009cd6e9 the version was
increased from 0.20.2 to 0.20.3 in the `.drone/drone.jsonnet` file, but
not in the `.drone/drone.yml`, which caused the image not to be built.

In order to trigger the build of 0.20.3, contents of the
`loki-build-image/` folder need to change, therefore the comment in the
Dockerfile.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>