fix: allow promtail docker_sd_config to handle long log lines
In promtail's docker_sd_config, parsing of a container's
logs would fail if any log line was over 64k bytes in length.
This is because bufio.Scanner uses a 64k buffer by default.
Since there doesn't appear to be a way to elegantly catch
`bufio.ErrTooLong` and carry on, the parsing approach was changed
to bufio.Reader which can handle lines of arbitrary length.