Introduce an "enable_groupcache" config to control initialization better (#6673)
* Introduce an "enable_groupcache" config to control initialization better

We initialize groupcache as a module, which happens after the config is parsed. A function named applyFIFOCacheConfig is called to enable fifocache if no cache is configured, and at the time of its calling the modules have not been initialized. This new "enable_groupcache" value is set if groupcache is configured in common, or this value is set manually. With this setting, we can prevent automatic fifocache settings.

Signed-off-by: Danny Kopping <danny.kopping@grafana.com>

* Add tests

Signed-off-by: Danny Kopping <danny.kopping@grafana.com>

* Simplifying condition

Signed-off-by: Danny Kopping <danny.kopping@grafana.com>