Add function to reset boltDBIndexClientWithShipper singleton
In a downstream project I have a test case that creates multiple stores
with different BoltDB shipper schema configurations. Since the client
for the store is a singleton and is re-used, there needs to be a
function to reset it as well.

This function must only be used in tests.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>