Fix free port allocation flake (#6898)
Occassionaly we can see test flakes like that:

> server: listen tcp :34473: bind: address already in use

This is because we allocated ports by checking if they are free and then
using them. Altough there is a chance that already allocated ports are
used twice. This changes prevents this from happening by keeping track
of the ports that are about to be used.