Bring helm chart back into Loki repo (#6829)
**What this PR does / why we need it**:

This is the first step of the process to bring the helm chart back into the Loki repo. We currently have ~7 public helm charts. This creates a lot of confusion for the community on which ones to start with, and is also a lot of additional surface area that is hard for Loki contributors to keep track of. By bringing the helm chart that we use ourselves back into the Loki repo, it will be a more clear starting point for helm users. It will also increase visibility for Loki contributors into helm PRs and issues. This will _not_ affect where the helm chart is published. It will still be published to the `grafana/helm-charts` following the same pattern as the Mimir helm chart.

The helm chart that we are currently using and most actively supporting is the `loki-simple-scalable` chart. Part of this effort is to rename that chart to just `loki`, which means it needs to be expanded to support the single binary deployment topology currently deployed by the existing `loki` helm chart. This PR is the first attempt to merge those two charts.

This chart currently works for a basic install of each deployment topology. There is more work to be done before we publish it. This work will include:

  * Making sure there is a migration path from the existing `loki` chart to this new one
  * Making sure there is a migration path from the existing `loki-simple-scalable` chart to this new one
  * Making sure the single binary deployment works in both `filesystem` and minio-backed multi-replica mode
  * Migrating CI from `grafana/helm-charts` for testing helm chart changes
  * Integrating with the github action in the `grafana/helm-charts` repo (this will publish it, so we will tackle this last)

A question I have for reviewers is what we should name each deployment mode (single binary vs. simple scalable)? We have received feedback from the community that some did not think "simple scalable" was suitable for production loads since it had "simple" in the name. This might be an opportunity to adjust that perception. Some options I've thought of include:

  * `single-binary` and `simple-scalable` (currently how it's coded)
  * `basic` and `scalable` (leaving open the option for `advanced` if we ever choose to support micro-services topology in this chart as well)
  * `ha`, `redundant` or similar vs. not

Fixes #6873