Adapt queryTimeout to be a per-tenant configuration (#6835)
What this PR does / why we need it:

* Add a new per-tenant query timeout
* Adds new middleware to query calls. This new middleware will timeout requests based on the new per-tenant query timeout
* Add span logs to query calls missing it
* Deprecate `engine.timeout` configuration

The motivation to change this configuration to be per-tenant instead of global is that it makes sense to alleviate the timeout for particular tenants or make it more strict for others. Especially useful for scenarios where the Loki client doesn't handle context cancellations correctly; in such scenarios, since Loki would still process expensive queries that were canceled, this allows one to have small timeouts for most tenants, which will help mitigate unnecessary work without having a timeout that is too short for important tenants.